Finished row col map
Only finds preimages of gcps which are fully in the image region of the base tiling. Moved functions out from parameter for checking this. When gcps are in cells in the base tiling which don't have a preimage in the parameter they will return a gcp in cells which don't exist in the parameter. The function for mapping them checks for this now.