Setup and start tidying (#7)
* start setup file, and making work with CayleyPerms

* black formatting

* some typing and repr

* motzkin file works now, but slow!

* tidy up delete rows and cols, and fuse params

* Start tidying files

Moved everything to 'old' folder, created new folder for mappling and parameter classes and new playground folder

* setup file

* Mapplings as a combinatorial class

Note: I added a function has_parameters to the mapped tiling class for the is_atom function. This checks if there are any avoiding or containing parameters and doesn't check for enumerating parameters

* mypy and pylint checks

from_dict function needs to be implemented for parameter class.
Other checks failed but I'm not sure how to fix them.

* update .gitignore

* Remove binary files

* Update setup.py description

Are the authors names correct??

* Update .gitignore

* added tox file

* add a workflow

* Update setup.py

* Delete playground directory

---------

Co-authored-by: Christian Bean <christianbean93@gmail.com>