Built New Cleaner Classes
An attempt to make the cleaner classes separate from base classes, and to make it possible to avoid keys. This assumes we won't need cleaner functions in any of our base classes

The register can be used as a function map like before, but also gives relevant functions an index attribute that we can use to sort them.

For both the function map and the indexing, a function must be decorated with @reg(idx)