Cleaner updates (#18)
* Added Missing Cleaning Functions

* A Basic Redundancy Check

* Update cleaner.py

* Test Adjustments

* Fixed repr methods

* Small Fixes

Fixed some bugs that I found while testing, and got rid of a print.

* Enumerator Flag

Added the ability to flag ParamCleaner methods that shouldn't be run on enumerators and added functionality for any future flags we may want.

Added a function to mapped tiling that runs a function only on the containers and avoiders

* Aesthetics

Changed the names of some register functions and fixed the register str and repr functions

* Debug and Loop

Added debugging to cleaners and began working on the loop cleaner

* Loop Cleaner

Simplified the loop cleaner. Made loop cleaning the default cleaner behavior.

* Direction Bug

Fixed the direction bug.

* Removed Requests

Got rid of the requests from earlier loop cleaner versions.

* Bugs on Bugs

Bad things happened when this file got merged with main, most should be fixed.

* Missed One

Missed one of the duplicate lines that came from merging main into this branch