Validate Positive Cells (#29)
* Make The Positive Cell Check

Added a function to parameter that checks if the positive cells are valid in regards to a base tiling and made a cleaning function that does this for all parameters (not enumerating) of a mapped tiling.

* Cleaner Exclusion

I made it so broken cleaning functions won't be added to the full cleanup method.

* Comprehension

I didn't need to do this at all.

* Test for Positive Cell Validation

Added a test for the new function

* Bug fix 1

Forget to get rid of code for positive rows

* Doc String

* Aesthtic Fixes

Made it pretty and good