merge main into strategy pack (#31)
* Point placement (#26)

* point placement

need to write update map functions

* maps for point placement

Need to write unfuse params function and check other functions work

* Finished param placement

Just need to put tests in

* Added tests

* fixing mypy

* pylint fixes

* more fixes

* more mypy errors

* better fix

* Final fix!

(sorry I can't see the errors on my laptop)

* fixed an error

* do not place in empty cells

---------

Co-authored-by: Christian Bean <christianbean93@gmail.com>

* Validate Positive Cells (#29)

* Make The Positive Cell Check

Added a function to parameter that checks if the positive cells are valid in regards to a base tiling and made a cleaning function that does this for all parameters (not enumerating) of a mapped tiling.

* Cleaner Exclusion

I made it so broken cleaning functions won't be added to the full cleanup method.

* Comprehension

I didn't need to do this at all.

* Test for Positive Cell Validation

Added a test for the new function

* Bug fix 1

Forget to get rid of code for positive rows

* Doc String

* Aesthtic Fixes

Made it pretty and good

* Reduce Parameter Gcps (#28)

* Reduce GCP Changes

Fixed the logic of the reduce parameter GCP function to be correct.

* Instantiated the Simplify Class

requirement_implied_by_some_requirements is not a static method

* Enabled Tests

No longer skipping the tests

* Fixed The Test File

Looks like I don't need to import pytest?

* use remove factors from obs with backmapped reqs

---------

Co-authored-by: Christian Bean <christianbean93@gmail.com>

* Row col sep fix (#27)

* point placement

need to write update map functions

* maps for point placement

Need to write unfuse params function and check other functions work

* Finished param placement

Just need to put tests in

* Added tests

* fixing mypy

* pylint fixes

* more fixes

* more mypy errors

* better fix

* Final fix!

(sorry I can't see the errors on my laptop)

* fixed an error

* do not place in empty cells

* the point placements pack and a test file called motzkin

* check rowcol sep doesn't change order of positive cells in parameters

* Updated row_col sep algorithm

Makes sure the relative order of positive cells in parameters is preserved, otherwise makes parameters empty.
Also added tests and removed a test for a cleaning function which isn't working yet.

* Fixed pylint

* Combinations in function and cleaner functions not all in register

* compare cells in both directions

* deleted staticmethod when dealing with conflicts

---------

Co-authored-by: Christian Bean <christianbean93@gmail.com>

---------

Co-authored-by: Abigail Ollson <145692399+Ollson2921@users.noreply.github.com>
Co-authored-by: Reed Acton <110858904+ReedActon@users.noreply.github.com>